On branch edburns-msft-jea-102-optional modified:   specification/src/main/asciidoc/platform/ApplicationProgrammingInterface.adoc
See https://dev.azure.com/jakarta-ee-azdo/jakarta-ee-azdo/_workitems/edit/105

Added to agenda for 2023-12-12:

Jea-102-optional

Jea-105-the-big-table: What to do?

Consider this big table from the platform spec.

Based on my understanding of the ruling of the spec committee as described in jea-101 the job to be done in JEA-105 is to

Remove all rows with OPT.

Fix the header text so it does not mention any of this REQ, POPT, OPT business.

Remove the Status column, since everything is required.

Signed-off-by: Ed Burns <Edward.Burns@microsoft.com>